% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_approximation.R
\name{.FUN_overall_sum_threshold}
\alias{.FUN_overall_sum_threshold}
\title{Combine sampled values from fire component distributions into overall response values}
\usage{
.FUN_overall_sum_threshold(
  xcomponents,
  weights = c(1, 1, 1),
  zero_threshold = 0.05
)
}
\arguments{
\item{xcomponents}{A three-column matrix or data frame of sample values for
fire regime components. Column order must be: frequency, severity, tsf.}

\item{weights}{A vector of three, positive weight values to apply to the
component variables when calculating the overall response as a weighted
sum. Order of elements is frequency, severity, tsf. Zero values are allowed
(might be useful for exploratory purposes). The default is to weight all
components equally. The provided values will be scaled to sum to 1.0.}

\item{zero_threshold}{Threshold value for individual component samples. If
any component has a lower value the overall response is set to zero.}
}
\value{
A vector of overall response values.
}
\description{
This function can be used with \code{\link{find_zoabeta_approximation}} to
calculate overall response values based on vectors of values sampled from the
fire component distributions. Overall response values are calculated using a
heuristic adopted by DPE researchers in which overall response is set to zero
if any component sample value is less than 0.05; otherwise the overall response
is calculated as the weighted sum of component values.
}
