% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{draw_response_curves}
\alias{draw_response_curves}
\title{Draw response curves and bounding regions for a given group}
\usage{
draw_response_curves(the_group)
}
\arguments{
\item{the_group}{Integer group number: a single value between 1 and the
number of groups defined in \code{\link{GroupExpertData}} (currently 18).}
}
\value{
A \code{ggplot} object.
}
\description{
This is a convenience function to graph the response of a specified group to
each of the three component variables of fire regime: Frequency (frequency of
fires within a preceding fifty year period), fire), Severity (severity of the
most recent fire expressed on a 1-8 scale) and TSF (time since last fire in
years). For each component, the response curve is drawn with a bounding
region, representing the point-wise degree of expert certainty at each
discrete value.
}
\details{
For each record in the data frame, the lower, most likely and upper estimates
for relative suitability are presently interpreted as the minimum, modal and
maximum parameters of a triangular distribution.
}
\examples{
library(fireresponse)

# Set a preferred theme for the graph appearance
library(ggplot2)
theme_set(theme_bw())

# Graph the response curves for group 3
draw_response_curves(3)

# Change the default labels and add a title using the `labs` function
# from the ggplot2 package
draw_response_curves(3) +
  labs(x = "Value", title = "Functional group 3")

}
