% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_prediction.R
\name{expert_to_fesm_severity}
\alias{expert_to_fesm_severity}
\title{Convert integer fire severity values on the scale used for expert-elicited
data to FESM severity values}
\usage{
expert_to_fesm_severity(sev, intermediate = c("error", "up", "down"))
}
\arguments{
\item{sev}{Vector of integer values for fire severity on the 0-8 scale used
for expert-elicited data.}

\item{intermediate}{A character value that controls how expert-scale values
\code{(1,3,5,7)} that map to intermediate values on the FESM scale should
be handled. Options are \code{'error'} (default) to stop with an error if
the input vector contains any such values; \code{'up'} to map these values
to the higher corresponding FESM values \code{(2,3,4,5)}; or \code{'down'}
to map these values to the lower corresponding FESM values
\code{(0,2,3,4)}. May be abbreviated.}
}
\value{
A vector of integer severity values on the FESM scale (0-5).
}
\description{
In the expert-elicited data for functional group responses, fire severity
values are coded on a 0-8 scale. This contrasts with the commonly used FESM
scale where severity is coded as: 0 (unburnt); 2 (low); 3 (moderate); 4
(high); 5 (extreme). Expert scale values \code{(0, 2, 4, 6, 8)} map directly
to FESM values \code{(0, 2, 3, 4, 5)}. The remaining expert-scale values
represent intermediate severities: 1 is unburnt/low (between FESM 0 and 2); 3
is low/moderate (between FESM 2 and 3); 5 is moderate/high (between FESM 3
and 4); and 7 is high/extreme (between FESM 4 and 5). The \code{intermediate}
argument can be used to control to handle these values, with options to
return the higher or lower FESM value, or (by default) to stop with an error
if any such values are encountered in the input vector.
}
\details{
Values of 1 (referred to as 'reserved class' in FESM documentation) do not
usually occur in FESM data and will never be returned by this function.
}
\examples{
# Expert severity codes that map directly to FESM severity codes.
expert_to_fesm_severity(c(0, 2, 4, 6, 8))  # returns c(0, 2, 3, 4, 5)

# Allow intermediate expert values and map each to the lower
# corresponding FESM value
expert_to_fesm_severity(c(1, 3, 5, 7), intermediate = "down")

# Map intermediate values to the upper corresponding FESM values
expert_to_fesm_severity(c(1, 3, 5, 7), intermediate = "up")

# If the argument intermediate is not 'up' or 'down', any intermediate
# values in the input vector will cause the function to stop with
# an error message
\dontrun{
expert_to_fesm_severity(3)
}

}
