% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\title{Simple graphical interface to explore group response curves}
\usage{
explorer()
}
\description{
This function launches a simple graphical interface to explore fire response
curves. Given a selected group, the response to each fire regime component
(frequency, severity and time since fire) based on expert-elicited data are
shown, together with the resulting overall response distribution(s). Overall
responses can be displayed for multiple fire regimes, but at present these
can only differ in time since and must share the same frequency and severity
values. This restriction will be relaxed in a future version of the package.
}
\details{
The \code{shiny} package must be installed to use this function.
}
