% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_approximation.R
\name{find_zibeta_approximation}
\alias{find_zibeta_approximation}
\title{Derive overall fire response and approximating zero-inflated beta distribution}
\usage{
find_zibeta_approximation(
  the_group,
  frequency,
  severity,
  tsf,
  nsamples = 10000,
  FUN = .FUN_OVERALL,
  pzero_threshold = 0.99
)
}
\arguments{
\item{the_group}{Integer group number: a single value between 1 and the
number of groups defined in \code{\link{GroupExpertData}} (currently 18).}

\item{frequency}{Fire frequency in the preceding fifty year period.}

\item{severity}{Severity of most recent fire.}

\item{tsf}{Time since last fire (years).}

\item{nsamples}{Number of random samples to generate from each component
triangular distribution. Default value is 10,000. Minimum allowable value
is 1000.}

\item{FUN}{A function that will be applied to the matrix of component sample
values to calculate the corresponding overall response values. Default is
the function \link{.FUN_overall_sum_threshold}.}

\item{pzero_threshold}{Maximum value of pzero (default=0.99) for which beta
parameters will be fitted.}
}
\value{
A numeric vector with three named elements:
\describe{
  \item{pzero}{probability of zero relative abundance value}
  \item{shape1}{first beta parameter for non-zero values (NA if pzero is close to 1.0)}
  \item{shape2}{second beta parameter for non-zero values (NA if pzero is close to 1.0)}
}
}
\description{
This function is mainly intended to generate the package data frame
\code{\link{GroupOverallResponse}} but might be useful for other purposes.
}
