% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_ggdata_zibeta}
\alias{get_ggdata_zibeta}
\title{Generate data to draw zero-inflated beta distributions with ggplot}
\usage{
get_ggdata_zibeta(pzero, shape1, shape2, type = c("density", "probability"))
}
\arguments{
\item{pzero}{One or more values for probability of zero.}

\item{shape1}{One or more values for the first beta parameter.}

\item{shape2}{One or more values for the second beta parameter.}

\item{type}{A character string specifying the type of y variable; either
'density' (default) or 'probability'. May be abbreviated.}
}
\value{
A data frame with columns: pzero, shape1, shape2, x, y and label. The
  label column is a factor that identifies the distribution for each data
  row. This can be used to facet or colour ggplot elements.
}
\description{
This function takes parameters for one or more zero-inflated beta
distributions and returns a data frame of density or probability values over
the unit interval. You can use it to explore the shape of alternative
distributions. If any of the arguments \code{pzero, shape1, shape2} have more
than one element, data is generated for all combinations.
}
\examples{
library(ggplot2)

# Get data for three distributions that differ in the
# probability of zero.
#
dat <- get_ggdata_zibeta(pzero = c(0, 0.3, 0.6), shape1 = 20, shape2 = 30)

ggplot(data = dat) +
  geom_line(aes(x=x, y=y, colour=label), linewidth=1.5) +
  scale_colour_viridis_d(end=0.8, direction = -1)

}
