% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_helper_functions.R
\name{get_group_ids}
\alias{get_group_ids}
\title{Get a vector of all integer ID values defined for functional groups in the expert data}
\usage{
get_group_ids()
}
\description{
This is a convenience function to get a vector of all functional group ID
values in ascending order.
}
