% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{get_zibeta_parameters}
\alias{get_zibeta_parameters}
\title{Get parameters for ZIBeta distributions that approximate group overall
response to specified fire regimes}
\usage{
get_zibeta_parameters(
  grp,
  frequency,
  severity,
  tsf,
  cross = TRUE,
  interpolation = TRUE
)
}
\arguments{
\item{grp}{One or more integer identifiers for group. Each value must be
between 1 and the number of groups defined in the
\code{GroupOverallResponse} table (currently 18).}

\item{frequency}{Integer vector with one or more values of fire frequency in
the preceding fifty year period.}

\item{severity}{Integer vector with one or more values of severity for the
most recent fire.}

\item{tsf}{Integer vector with one or more values of time since last fire
(years).}

\item{cross}{Logical value. If \code{TRUE} (default), a set of fire regimes
will be created for all combinations of the fire component vectors:
frequency, severity and tsf values. In this case, vectors can have
different lengths; e.g. 2 frequency values x 3 severity values x 2 tsf
values will result in 12 fire regimes. If \code{FALSE}, a fire regime
\emph{i} is defined by the \emph{i}th element of each of the three
component vectors. In this case, the vectors must all have the same length.}

\item{interpolation}{If \code{TRUE} (default), interpolation will be performed
for any combinations of frequency, severity and time since fire values that
do not appear in the \code{GroupOverallResponse} look-up table. If
\code{FALSE}, a warning message will be issued for such combinations and no
data returned for them.}
}
\value{
A data frame suitable for use with ggplot, with columns:
  group; pzero; shape1; shape2.
}
\description{
Given a set of integer group IDs, and one or more fire regimes defined as
combinations of frequency, severity and time since fire values, this function
returns the parameters for zero-inflated beta distributions that approximate
the group's overall response to each regime. If a regime is defined in the
package look-up table \code{\link{GroupOverallResponse}}, the function simply
returns the corresponding distribution parameters. For other fire regimes,
the parameters will be derived as the weighted mixture of beta distributions
for the most similar reference fire regimes.
}
\details{
At the moment, interpolation between fire regimes only involves time since
fire, since all integer values for frequency and severity between zero and
the maximum values considered during expert-elicitation are represented in
the look-up table. Extrapolation beyond the range of any of the three fire
regime component variables is not supported.
}
\seealso{
\code{\link{draw_overall_response}} \code{\link{get_overall_response_data}}
}
