% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibeta_functions.R
\name{qzibeta}
\alias{qzibeta}
\title{Distribution function for the zero-inflated beta distribution}
\usage{
qzibeta(p, pzero, shape1, shape2, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{pzero}{Probability of zero value.}

\item{shape1, shape2}{Non-negative parameters of the beta distribution for
non-zero values.}

\item{lower.tail}{If \code{TRUE} (default), probabilities are \deqn{P[X \le x]},
otherwise \deqn{P[X \gt x]}.}

\item{log.p}{If \code{TRUE}, log probability values are returned.}

\item{x}{Vector of probabilities.}
}
\value{
Vector of probability values.
}
\description{
Distribution function for the zero-inflated beta distribution
}
