% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibeta_functions.R
\name{rzibeta}
\alias{rzibeta}
\title{Draw random values from a zero-inflated beta distribution}
\usage{
rzibeta(n, pzero, shape1, shape2)
}
\arguments{
\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}

\item{shape1, shape2}{Non-negative parameters of the beta distribution for
non-zero values.}
}
\value{
Vector of random values.
}
\description{
Draw random values from a zero-inflated beta distribution
}
