% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_approximation.R
\name{.FUN_overall_product}
\alias{.FUN_overall_product}
\title{Combine sampled values from fire component distributions into overall response values}
\usage{
.FUN_overall_product(xcomponents, weights = c(1, 1, 1), lambda = 1/3)
}
\arguments{
\item{xcomponents}{A three-column matrix or data frame of sample values for
fire regime components. Column order must be: frequency, severity, tsf.}

\item{weights}{A vector of three, positive weight values to apply to the
component variables when calculating the overall response as a weighted
product. Order of elements is frequency, severity, tsf. Input weights will
be scaled so that they sum to 3.0. Zero values are allowed (might be useful
for exploratory purposes). The default is to weight all components equally.}

\item{lambda}{A single numeric value (default = 1/3) to map the weighted
product to the overall response value as:
\deqn{y_{overall} = (\sum{x_i wt_i})^\lambda }}
}
\value{
A vector of overall response values.
}
\description{
This function can be used with \code{\link{find_zibeta_approximation}} to
calculate overall response values based on vectors of values sampled from the
fire component distributions. It combines the component values as a weighted
product raised to a power \code{lambda} (default = 1/3).
}
