% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{draw_overall_response}
\alias{draw_overall_response}
\title{Graph group overall response to fire regime(s)}
\usage{
draw_overall_response(
  the_group,
  frequency,
  severity,
  tsf,
  cross = TRUE,
  linewidth = 1,
  draw_pzero = TRUE,
  colour_map = "viridis"
)
}
\arguments{
\item{the_group}{One or more integer identifiers for group. Each value must
be between 1 and the number of groups defined in
\code{\link{GroupExpertData}} (currently 18).}

\item{frequency}{Integer vector with one or more values of fire frequency in
the preceding fifty year period.}

\item{severity}{Integer vector with one or more values of severity for the
most recent fire.}

\item{tsf}{Integer vector with one or more values of time since last fire
(years).}

\item{cross}{Logical value. If \code{TRUE} (default), a set of fire regimes
will be created for all combinations of the fire component vectors:
frequency, severity and tsf values. In this case, vectors can have
different lengths; e.g. 2 frequency values x 3 severity values x 2 tsf
values will result in 12 fire regimes. If \code{FALSE}, a fire regime
\emph{i} is defined by the \emph{i}th element of each of the three
component vectors. In this case, the vectors must all have the same length.}

\item{linewidth}{Value to pass to ggplot for line width (default = 1).}

\item{draw_pzero}{Logical. If \code{TRUE} (default), a separate graph showing
the probability of zero relative abundance under each fire regime is drawn
and placed above the graph of overall response distribution using the
\code{patchwork} package.}

\item{colour_map}{Character string indicating the colour map option to use
with the viridis palette. Default is \code{"cividis"}. See
\code{\link[ggplot2]{scale_colour_viridis_d}} for details.}
}
\value{
A ggplot object.
}
\description{
Presently, this just graphs the beta distribution that approximates the
overall group response for a given fire regime. If multiple groups and/or
fire regimes are specified, the graph for each group is displayed as a
separate facet with the curve for each fire regime identified by colour.
}
