% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{draw_overall_response}
\alias{draw_overall_response}
\title{Graph group overall response to fire regime(s)}
\usage{
draw_overall_response(
  the_group,
  frequency,
  severity,
  tsf,
  linewidth = 1,
  draw_pzero = TRUE,
  colour_map = "cividis"
)
}
\arguments{
\item{the_group}{One or more integer identifiers for group. Each value must
be between 1 and the number of groups defined in
\code{\link{GroupExpertData}} (currently 18).}

\item{frequency}{Integer vector with one or more values of fire frequency in
the preceding fifty year period.}

\item{severity}{Integer vector with one or more values of severity for the
most recent fire.}

\item{tsf}{Integer vector with one or more values of time since last fire
(years).}
}
\value{
A ggplot object.
}
\description{
Presently, this just graphs the beta distribution that approximates the
overall group response for a given fire regime. If multiple groups and/or
fire regimes are specified, the graph for each group is displayed as a
separate facet with the curve for each fire regime identified by colour.
}
