% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{draw_response_curves}
\alias{draw_response_curves}
\title{Draw response curves and bounding regions for a given group}
\usage{
draw_response_curves(the_group)
}
\arguments{
\item{the_group}{Integer group number: a single value between 1 and the
number of groups defined in \code{\link{GroupExpertData}} (currently 18).}
}
\value{
A \code{ggplot} object.
}
\description{
This is a convenience function to graph the response of a specified group to
each of the three component variables of fire regime: TSF (time since last
fire), severity of the most recent fire and frequency of fires within a
preceding fifty year period. For each component, the response curve is drawn
with a bounding region, representing the point-wise degree of expert
certainty at each discrete value.
}
\details{
For each record in the data frame, the lower, most likely and upper estimates
for relative abundance are presently interpreted as the minimum, modal and
maximum parameters of a triangular distribution.
}
\examples{
# Graph the response curves for group 3
draw_response_curves(3)

# Change the default labels and add a title
library(ggplot2)
draw_response_curves(3) +
  labs(x = "Value", title = "Functional group 3")

}
