% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibeta_functions.R
\name{dzibeta}
\alias{dzibeta}
\title{Density function for the zero-inflated beta distribution}
\usage{
dzibeta(x, pzero, shape1, shape2, log = FALSE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{pzero}{Probability of zero value.}

\item{shape1, shape2}{Non-negative parameters of the beta distribution for non-zero values.}

\item{log}{If \code{TRUE}, log density values are returned.}
}
\value{
Vector of density values.
}
\description{
Density function for the zero-inflated beta distribution
}
