% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{get_overall_response_data}
\alias{get_overall_response_data}
\title{Prepare data to graph group overall response to fire regime(s)}
\usage{
get_overall_response_data(the_group, frequency, severity, tsf, dx = 0.005)
}
\arguments{
\item{the_group}{One or more integer identifiers for group. Each value must
be between 1 and the number of groups defined in
\code{\link{GroupExpertData}} (currently 18).}

\item{frequency}{Integer vector with one or more values of fire frequency in
the preceding fifty year period.}

\item{severity}{Integer vector with one or more values of severity for the
most recent fire.}

\item{tsf}{Integer vector with one or more values of time since last fire
(years).}
}
\value{
A data frame.
}
\description{
Prepare data to graph group overall response to fire regime(s)
}
