% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{get_overall_response_data}
\alias{get_overall_response_data}
\title{Prepare data to graph group overall response to fire regime(s)}
\usage{
get_overall_response_data(
  grp,
  frequency,
  severity,
  tsf,
  cross = TRUE,
  interpolation = TRUE,
  dx = 0.005
)
}
\arguments{
\item{grp}{One or more integer identifiers for group. Each value must
be between 1 and the number of groups defined in the \code{GroupOverallResponse}
table (currently 18).}

\item{frequency}{Integer vector with one or more values of fire frequency in
the preceding fifty year period.}

\item{severity}{Integer vector with one or more values of severity for the
most recent fire.}

\item{tsf}{Integer vector with one or more values of time since last fire
(years).}

\item{cross}{Logical value. If \code{TRUE} (default), a set of fire regimes
will be created for all combinations of the fire component vectors:
frequency, severity and tsf values. In this case, vectors can have
different lengths; e.g. 2 frequency values x 3 severity values x 2 tsf
values will result in 12 fire regimes. If \code{FALSE}, a fire regime
\emph{i} is defined by the \emph{i}th element of each of the three
component vectors. In this case, the vectors must all have the same length.}

\item{interpolation}{If \code{TRUE} (default), interpolation will be performed
for any combinations of frequency, severity and time since fire values that
do not appear in the \code{GroupOverallResponse} look-up table. If
\code{FALSE}, a warning message will be issued for such combinations and no
data returned for them.}

\item{dx}{Increment for relative abundance values.}
}
\value{
A data frame suitable for use with ggplot, with columns:
  group; frequency; severity; tsf; relabund; density; density_adjusted;
  regime.
}
\description{
This is a convenience function that creates a data frame suitable for
graphing the distribution of expected relative abundance values for one or
more fire regimes. It is mainly intended as a helper function to be called by
the \code{\link{draw_overall_response}} function, but can also be called
explicitly.
}
\details{
Where a requested combination of frequency, severity and time since fire does
not exist in the \code{GroupOverallResponse} look-up table, the overall
response will be derived as the weighted mixture of beta distributions for
the nearest defined combinations. At the moment, this interpolation between
values is relevant to time since fire, since all integer values for frequency
and severity between zero and the maximum values considered during
expert-elicitation are represented in the look-up table.
}
\seealso{
\code{\link{draw_overall_response}}
}
