% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_approximation.R
\name{get_tri_samples}
\alias{get_tri_samples}
\title{Sample component triangular distributions for a group and fire regime}
\usage{
get_tri_samples(the_group, frequency, severity, tsf, nsamples = 10000)
}
\arguments{
\item{the_group}{Integer group number: a single value between 1 and the
number of groups defined in \code{\link{GroupExpertData}} (currently 18).}

\item{frequency}{Fire frequency in the preceding fifty year period.}

\item{severity}{Severity of most recent fire.}

\item{tsf}{Time since last fire (years).}

\item{nsamples}{Number of random samples to generate from each component
triangular distribution.}
}
\value{
A data frame with \code{nsamples} rows and a column for each of
  frequency, severity and tsf.
}
\description{
Note: This is a private function and is mainly intended for generating the
package data frame \code{\link{GroupOverallResponse}}
}
\keyword{internal}
