% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GroupExpertData}
\alias{GroupExpertData}
\title{Expert-elicited group responses to fire regime variables}
\format{
A data frame with 666 rows and 6 columns:
\describe{
\item{group}{Integer identifier for the functional group (1-18)}
\item{type}{Type of fire variable, coded as a factor with levels: frequency; severity; tsf}
\item{value}{Value of the given fire variable (e.g. years for TSF)}
\item{ra_lwr}{Estimate of lowest expected relative abundance}
\item{ra_mode}{Estimate of most likely value for relative abundance}
\item{ra_max}{Estimate of highest expected relative abundance}
}
}
\usage{
GroupExpertData
}
\description{
A data frame with, for each functional group, bounded estimates for the
expected relative abundance (on the unit interval) given discrete values for
time since last fire (years); severity of last fire; and frequency of fire in
the preceding fifty year period.
}
\keyword{datasets}
